refactor(sidebar): move resolveTopicActiveContext to shared, stabilize topic arrays
Addresses reviewer feedback on #1068:

- Move resolveTopicActiveContext into shared/workspace-topics.ts alongside
  the other pure topic utilities so it is reusable and testable without
  importing the React component.
- Memoize the topics/surfaces/search-results arrays passed to
  TopicSidebarView with stable empty-array constants, so its model and
  topicsById memoization is no longer invalidated on every parent render.
- Drop the now-unused eslint-disable complexity directive.

Refs #1061
